requirements/base.txt
# base.txt is a pip requirements file which describes the necessary
# distributions required to run this project. If you need something that is
# project specific not listed here use project.txt. You can, of course, update
# versions of distributions here if needed.

--extra-index-url=http://dist.pinaxproject.com/dev/

Django==1.2.4
Pinax==0.9a1

django-admin-tools==0.4.0
django-debug-toolbar==0.8.3
django-extensions
django-filebrowser==3.1
django-grappelli==2.2
django-staticfiles==0.2.0

south

